--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -131,31 +131,6 @@ jobs:
           command: cargo test --workspace
       - cache_save
 
-  # Temporary test to see how much slower it is to test using full
-  # object store features
-  #
-  # I don't think this adds much actual coverage
-  test_object_store:
-    docker:
-      - image: quay.io/influxdb/rust:ci
-    resource_class: xlarge # use of a smaller executor tends crashes on link
-    environment:
-        # Disable incremental compilation to avoid overhead. We are not preserving these files anyway.
-        CARGO_INCREMENTAL: "0"
-        # Disable full debug symbol generation to speed up CI build
-        # "1" means line tables only, which is useful for panic tracebacks.
-        RUSTFLAGS: "-C debuginfo=1"
-        RUST_BACKTRACE: "1"
-    steps:
-      - checkout
-      - rust_components
-      - cache_restore
-      - run:
-          name: Cargo test with object store
-          command: cargo test --workspace --features="aws,gcp,azure"
-      - cache_save
-
-
   test_kafka_integration:
     machine: true
     resource_class: xlarge
@@ -331,7 +306,6 @@ workflows:
       - lint
       - protobuf-lint
       - test
-      - test_object_store
       - test_kafka_integration
       - test_influxdb2_client
       - build