--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -151,7 +151,7 @@ jobs:
       - rust_components
       - cache_restore
       - run:
-          name: Cargo test
+          name: Cargo test with object store
           command: cargo test --workspace --features="aws,gcp,azure"
       - cache_save