--- a/Cargo.toml
+++ b/Cargo.toml
@@ -131,3 +131,8 @@ rand = "0.8.3"
 rdkafka = "0.26.0"
 reqwest = "0.11"
 tempfile = "3.1.0"
+
+[features]
+azure = ["object_store/azure"]
+gcp = ["object_store/gcp"]
+aws = ["object_store/aws"]